Made some changes in RBFInterpolation.cs  and  NearestNeighborSearch.cs
